requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.1
torch==2.4.1
torchvision==0.19.1
timm==1.0.9
pillow>=10.0.0
numpy>=1.24.0
python-multipart>=0.0.9
ultralytics>=8.0.0
